requirements.txt
streamlit
google-generativeai
python-dotenv
firecrawl
chromadb
pandas
requests
beautifulsoup4
